Simple fix for bug #23017 (https://gna.org/bugs/?23017).
This is the multidimensional numpy arrays are not being stored as IEEE 754 arrays in the XML state
and results files.  The problem was a relatively recent regression caused by a change to the
is_float_matrix() function of the lib.arg_check module.  It was simply that the default dims keyword
argument value was changed from None to (3, 3).  Therefore any call to the function without
supplying the dims argument would fail if the matrix was not of the (3, 3) shape.