Change to the Structure.test_align system test.
The molecules argument for the structure.align user function has been changed to match the models
argument, in that it now needs to be a list of lists with the first dimension matching the pipes
argument.  This change is to help with the implementation of the new structure.align functionality.