Added the displace_id argument to the structure.align and structure.superimpose user functions.
This gives both of these user functions finer control over which atoms are translated and rotated by
the algorithm.  This allows, for example, to align structures based on a set of backbone heavy atoms
while the protons and side chains are displaced by default.  Or if a domain is aligned, then just
that domain can be displaced.