More robustness for the spin viewer GUI window update_*() methods.
The data returned from the self.tree.GetItemPyData(key) call can in rare racing cases not contain
the 'id' key.  This is now being checked for and are being skipped in the update_mol(), update_res()
and update_spin() methods.

The problem was found in the Mf.test_bug_20479_gui_final_pipe system test when running the command:
for i in {1..10}; do ./relax --gui-tests --time -d &>> gui_tests.log; done