Extended the user function 'monte_carlo.create_data', to allow for the defition of the STD to use in gauss distribution.
This is for creation of Monte-Carlo simulations, where one has perhaps gained information about the expected errors of the datapoints, which is not measured.

Task #7882 (https://gna.org/task/?7882): Implement Monte-Carlo simulation, where errors are generated with width of standard deviation or residuals.): Implement Monte-Carlo simulation, where errors are generated with width of standard deviation or residuals.