Modified the Needleman-Wunsch sequence alignment algorithm.
The previous attempt was buggy.  The algorithm has been modified to match the logic of the GPL
licenced EMBOSS software (http://emboss.sourceforge.net/) to allow for gap opening and extension
penalties, as well as end penalties.  No code was copied, rather the algorithm for creating the
scoring and penalty matrices, as well as the traceback matrix.