Large simplification of the atomic coordinate assembly code in the internal structural object.
This is in the lib.structure.internal.coordinates.assemble_coord_array() function.  The logic of the
function has recently changed due to the introduction of the pairwise sequence alignments.  This
caused a lot of code to now be redundant, and also incorrect in certain cases.  This simplification
fixes the problem caught by the Structure.test_align_molecules2 system test.