Created the Sequence.test_align_molecules system test.
This will be used to implement the sequence.align user function which will be used for performing
sequence alignments on structural data within the relax data store and storing the data in the data
pipe independent sequence_alignments data store object (which will be an instance of
data_store.seq_align.Sequence_alignments).  The system test also checks the XML saving and loading
of the ds.sequence_alignments data structure.