Renamed the Sequence.test_align_molecules system test to Structure.test_sequence_alignment_molecules.
As the sequence alignment is dependent on the structural data in the relax data store, the user
function for sequence alignment would be better named as structure.sequence_alignment.  The
sequence.align user function is not appropriate as all other sequence user functions relate to the
molecule, residue, and spin data structure of each data pipe rather than to the structural data.