Important formatting improvement for the description in the GUI user function windows.
Previously lists, item lists, and prompt items were spaced with one empty line at the top and two at
the bottom.  The two empty lines at the bottom was an accident caused by how the list text elements
were built up.  Now the final newline character is stripped so that the top and bottom of the lists
only consist of one empty line.  The change can give a lot more room in the GUI window.