The Structure.test_create_diff_tensor_pdb_oblate now uses oblate diffusion relaxation data.
This fixes bug #23232 (https://gna.org/bugs/?23232), the failure of this system test on Mac OS X.
The problem was that the system test was previously using relaxation data for prolate spheroidal
diffusion and fitting an oblate tensor to that data.  This caused the solution to be slightly
different on different CPUs, operating systems, Python versions, etc. and hence the PDB file
representation of the diffusion would be slightly different.