Big bug fix for the GUI tests on MS Windows systems.
On MS Windows systems, the GUI tests were unable to complete without crashing.  This is because each
GUI element requires one 'User object', and MS Windows has a maximum limit of 10,000 of these
objects.  The GUI tests were taking more than 10,000 and then Windows would say - relax, you die
now.

The solution is that after each GUI test, all user function windows are destroyed.  The user
function page is a wx.Panel object, so this requires a Destroy() call.  But the window is a Uf_page
instance which inherits from Wiz_page which inherits from wx.Dialog.  Calling Destroy() on MS
Windows and Linux works fine, but is fatal on Mac OS X systems.  So the solution is to call Close()
instead.