Simplification of the GUI analysis post_reset() method.
This now uses the delete_all() and hence delete_analysis() methods to clean up the GUI.  The reset
argument has been added to skip the manipulation of relax data store data, as the data store is
empty after a reset.  However the calling of the delete_analysis() method now ensures that the
analysis specific delete() method is now called so that the GUI elements can be properly destroyed.