Fix for GUI wizards and GUI user functions.
The recent memory management changes caused the wizard windows to have an incorrect layout so that
the wizard pages were not visible.  Reperforming a layout of the GUI elements did not help.  The
solution is to not initialise sets of max_pages of wx.BoxSizer elements in the wizard __init__()
method, but to generate and append these dynamically via the add_page() method.  The change now
means that there are no longer multiple unused wx.BoxSizer instances generated for each wizard
window created.