Implementation of the statistics.aic user function.
This is very similar to the statistics.model user function - the code was copied and only slightly
modified.  The new user function will calculate the current chi-squared value per model, obtain the
model statistics, calculate the AIC value per model, and store the AIC value, chi-squared value and
number of parameters in the appropriate location for the model in the relax data store.