Fix for bug #23259 (https://gna.org/bugs/?23259).
This is the broken user functions in the prompt UI with the RelaxError: The user function 'X' has
been renamed to 'Y'.  The problem was that the only the first part of the user function name, for
example 'minimise' from 'minimise.calculate' was being checked in the user function name translation
table.  As the minimise user function has been renamed to minimise.execute, 'minimise' is in the
translation table and hence minimise.calculate was being identified as the minimise user function.
Now the full user function name is reconstructed before checking the translation table.