Created the Test_object.test_add_atom_sort unit test.
This is from the _lib._structure._internal.test_object unit test module.  The test will be used to
implement the sorting of input data by residue number in the add_atom() internal structural object
method.  This will mean that added atoms will be placed in residue sequence order, so that output
PDB files are correctly ordered.