Implementation of methods for sorting sequence data in the internal structural object.
The information is sorted in the molecule container level using the new MolContainer._sort() private
method.  This uses the _sort_key() helper method which determines what the new order should be.
This is used as the 'key' argument for the Python sort() method.  Instead of list shuffling, new
lists in the correct order are created.  Although not memory efficient, this might be faster than
shuffling.