Implemented the back-end of the frame_order.distribute user function.
This follows the design of the pseudo-Brownian simulation frame_order.simulate user function.  The
specific_analyses.frame_order.uf.distribute() function has been created as a modified copy of the
simulate() function of the same module.  This simply performs checks and assembles the data, passing
into the new lib.frame_order.simulate.uniform_distribution() function, which itself is a modified
copy of the brownian() function in the same module.