Fix for bug #23294 (https://gna.org/bugs/?23294).
This is the automatic merging of PDB molecules resulting in an IndexError.  Now if only a single
molecule name is supplied, this will be used for all molecules in the PDB file.  The result is that
the structural data will all be automatically merged into a single molecule.  This merging is
communicated to the user via the current printouts.