Fix for bug #23295 (https://gna.org/bugs/?23295).
This is the PDB secondary structure HELIX and SHEET records not updating when merging molecules.
The problem was that the algorithm for changing the molecule numbers for the helix and sheet
metadata when calling the structure.read_pdb user function was far too simplistic.  Therefore the
logic has been completely rewritten.

Now the helix and sheet metadata are stored in temporary data structures in the _parse_pdb_ss()
method.  As the molecules are being read from the PDB records, new data structures containing the
original molecule numbers and new molecule numbers are created.  The helix and sheet metadata is
then stored in the internal structural object via the pack_structs() method, and the molecule
indices of the metadata changed based on the two molecule number remapping data structures.