Large speed up of the structure.web_of_motion user function.
With the introduction of the _sort() internal structural object method and it being called by the
add_atom(), the structure.web_of_motion user function was now painfully slow.  As sorting the
structural data is unnecessary for the backend of this user function, the add_atom() boolean
argument 'sort' has been added to turn the sorting on and off, and the structure.web_of_motion
backend now sets this to False.