Fix of a fatal bug preventing the frame order analysis to be run on a multi-processor system.
The multi-processor code was calling the count_sobol_points() function of the
specific_analyses.frame_order.optimisation module to give feedback when calling the minimise.execute
or minimise.calculate user functions.  However this was run in the slave command run() method, hence
would be executed on the slave.  The problem is that count_sobol_points() performs a number of
checks on the current data pipe, however the slaves do not have any data pipes set up.