Bug fix for the results.write user function when loading relax state files.
The results.write user function can load not only the results file consisting of a single data pipe,
but also relax state files if only a single pipe is present.  However this was causing the current
data pipe and other pipe-independent data (sequence alignments and the GUI) to be overwritten, just
as when loading a state file.  Now only the data from the data pipe will be loaded and the pipe
independent data in the state file will be ignored.