Big bug fix for the pipe_control.interatomic.interatomic_loop() function.
This was identified in the Rdc.test_rdc_copy_different_spins system test.  The problem was that the
pipe argument was being ignored when looking up the spin containers.  Hence if the pipe being worked
on was not the current data pipe, and the spin sequences were not identical, the function would
fail.  This mainly affects the rdc.copy user function.