Clarification of the RDC and PCS Q factors.
This affects the rdc.calc_q_factors and pcs.calc_q_factors user functions, as well as all other
operations involving the calculation of Q factors.  The printouts have been modified to clarify if
the normalisation is via the tensor size (2Da^2(4 + 3R)/5) or via the sum of data squared, and the
separation of the two is now clearer.  This allows for better RDC vs. PCS comparisons.

In addition, the data pipe variable names have been updated to reflect the normalisation, so it is
instantly known when looking at the XML contents of results or save files which was used.  The
backwards compatibility hooks have been modified to support the data pipe variable name changes.