Resized all fixed-sized GUI wizards to fit on 1024x768 pixel wide displays.
The problem was reported by Lora Picton in the thread starting at
http://thread.gmane.org/gmane.science.nmr.relax.user/1813 .

Both the spin loading wizard of the spin viewer window and the relaxation data loading wizard used
currently in the model-free analysis tab and BMRB export page were fixed.  These both had the
y-dimension set to 800 pixels, hence parts of the window would be out of view.