Wrote a method to store parameter data and dispersion curves, for the protocol of repeated analysis.
This is to prepare for analysis in other programs.
The method loops throug the data pipes, and write the data out.

It then write a bash script, that will concatenate the data in an matrix array style, for reading and processing in other programs.

Task #7826 (https://gna.org/task/?7826): Write an python class for the repeated analysis of dispersion data.