Added an exception system to the find_unused_imports.py developer script.
Sometimes pylint will give an "Unused import" warning for imports that are needed by the module.
Therefore an exception list of the file name and module has been created to skip these warnings.
The list covers the dep_check module and all of the profiling_*.py scripts in the directory
test_suite/shared_data/dispersion/profiling/.