Fixes for sub and superscripts throughout the manual.
This introduces {} around all sub and superscripted \textrm{} instances.  This is not needed for the
PDF version of the manual as the missing bracket problem is avoided, but it affects the HTML version
of the manual compiled by latex2html, which requires the correct notation.  The fixes are for both
the new frame order chapter as well as the relaxation dispersion chapter.