Created the Mf.test_bug_23933_relax_data_read_ids system test.
This is designed to catch bug #23933 (https://gna.org/bugs/?23933), the "NameError: global name
'ids' is not defined" problem when loading relaxation data.

A truncated version of the PDB file and relaxation data, the full versions of which are attached to
the bug report, consisting solely of residues 329, 330, and 331 have been added to the test suite
shared data directories, and the system test written to catch the NameError.