Fix for bug #23933 (https://gna.org/bugs/?23933).
This is the "NameError: global name 'ids' is not defined" problem when loading relaxation data.

The bug was introduced back in November 2014 (r26588), and is due to some incomplete error handling
code.  The problem is that the spin type that the relaxation data belongs to (@N vs. @H) has not
been specified.  Now the correct RelaxMultiSpinIDError is raised.  The 'ids' variable did not exist
- it was code that was planned to be added, but never was and was forgotten.