Partial implemented of the PCA analysis in the relax library.
This is for the new structure.pca user function.  The lib.structure.pca module has been created, and
the pca_analysis() function created to calculate the structure covariance matrix, via the
covariance() function, and then calculate the eigenvalues and eigenvectors of the covariance matrix,
sorting them and truncating to the desired number of PCA modes.