Fix for the Mf.test_bug_24131_missing_interaction system test.
This is part of bug #24131 (https://gna.org/bugs/?24131), the BMRB export failure with the
SpinContainer object having no s2 value.  The problem was when assembling the diffusion tensor data.
The spin_loop() function was being called, as the diffusion tensor is reported for all residues.
Therefore the skip_desel=True has been added to match the model-free part.