relax version 4.0.1.
This is a major feature and bugfix release.  Features include the new structure.pca user function
for performing a principle component analysis (PCA) of a set of structures, handling of replicated
R2eff data points in the dispersion analysis, improvements in the handling of PDB structures, the
protection against numpy >= 1.9 FutureWarnings for a number of soon to change behaviours in numpy,
and addition of a deployment script for the Google Cloud Computing.  Bugfixes include an error when
loading relaxation data, the CSA constant equation in the manual, missing information in the relax
state and results files, loading of certain state files in the GUI, running relax with no graphical
display and using matplotlib, BMRB export failure when a spin container is missing data or
parameters.