Added support for concatenating atomic positions in the structure.load_spins user function.
Together with the spin_num flag set to False, this allows for atomic positions to be read from
multiple homologous structures with different PDB atomic numbering.  The spin containers will be
created from the first structure, in which the spin is defined, and the atomic position from
subsequent structures will be appended to the list of current atomic positions.