Added some missing oxygen icons to allow the relax manual to compile.
These are the 128x128 EPS versions of the places/folder-development.png and
places/folder-favorites.png Oxygen icons recently introduced.  For completeness the 32x32, 48x48,
and 128x128 PNG versions of the icons have also been added.

To help create these EPS icons in the future, the graphics/README file has been added with a
description of the *.eps.gz file creation.