Huge improvement for the responsiveness of the relax GUI.
The relax controller window log panel was being updated with a wx.CallAfter() call after every
write to the IO streams.  If a relax analysis was proceeding very quickly, which is the case in
most analyses, this created a huge backlog of GUI updates.  The result was that the GUI would
freeze, running at 100% CPU usage in its own thread, with the analysis running at 100% on another
thread.

The fix was to shift the log panel write() call to be triggered by the Timer already being used by
the gauges, rather than by the IO stream write() methods.  The text was already placed on a Queue
object, so this change is very simple.

Another small change was made to the log panel write() method to avoid a number of unneeded wx
calls.  This should also have a significant impact on the GUI updating.