Bug fix by redesigning the GUI pipe editor pop up menu.
The menu now uses IDs to associate menu items with the correct method to call.  Previously all menu
entries were calling the method of the last menu entry, which was in most cases the pipe switching
method.  As the pipe deletion method is now properly exposed, the Question dialog was increased in
size to be able to see all the text.