Safe execution of all of the auto-analyses (those that acquire the execution lock).
The whole of the __init__() code of the auto-analyses is now wrapped within a try-finally set of
statements.  This is to be absolutely sure that the execution lock is released.  This is not always
the case, for example the Relax_fit.test_auto_analysis_pipe_name system test was not releasing the
lock due to a RelaxError, and this was causing the later GUI tests to fail.