Support for a spin hash fast lookup table for the molecule, residue and spin data structures.
The fast lookup table is stored as dp.mol._spin_hash_lookup.  This matches the
dp.mol._spin_id_lookup fast lookup table, but is a simpler table to maintain as there is only one
hash ever per spin and that hash is unique.  The table is maintained by the
pipe_control.mol_res_spin module.