Redesign of the create_spin() function of the pipe_control.mol_res_spin module.
This function is the backend of the spin.create user function and is also used throughout relax.
Instead of creating a single residue or spin, if only a name and not number is supplied, now
multiple spins are created.  If the residue name is supplied but not the residue number, now all
residues matching the given name will have new spins created.  For example creating the spin with
the name 'NE1' and only specifying the residue name 'TRP', then all tryptophans in all molecules
will have NE1 indole side-chain spins created.  This makes the operation of the spin.create user
function more logical for the user.