Split the FSF Copyright Validation script into a configuration file and an executable script.
The configuration part of the script has been retained but with all data stripped to be able to
provide a blank template for a new configuration file.  And the new mimetypes section has been
converted into a variable rather than manipulating the mimetypes Python module so that the
configuration script requires no Python imports.