Shifted the epydoc API documentation copyright notice insertion into the scons script.
This notice was previously hardcoded into the devel_scripts/google_analytics.js script - as that is
the GPLv3+ copyright notice of that script with the date of 2012.  Instead the copyright notice in the
Google analytics script is now skipped and the correct FDLv1.3 copyright notice with the current
year programmatically inserted via the scons/manuals.py script.