Added the MIT licence with copyright notices to the top of all files.
The origin of all code was traced back through the MATLAB sources, FORTRAN90 sources, and FORTRAN77
sources.  The original f77 code did not contain any shared lines of code with the f90 code, so no
copyright statements for Bennett Fox were added.  Comments were added to each function to document
the history of all of the code.