FSF Copyright Validation script: Support for tracking files renamed in later repositories.
In this case, a file rename in the current git repository would not allow the file to be found in
the SVN archive repository.  The history of the later repository is now used to find all file
renames after the end of the earlier repository.  False git history is also correctly handled.