Adding sending mails to nmr-relax-devel@lists.sourceforge.net.
This introduces a spamming problem.

Everyone who forks this project and have travis setup for their user
will spam the develop mailing list.

To limit this, there are options in travis:
 * https://docs.travis-ci.com/user/notifications/
 * https://docs.travis-ci.com/user/conditional-builds-stages-jobs

Introducing a condition like
  if: branch = master

seems not to be implemented yet.
 * https://github.com/travis-ci/travis-ci/issues/1405

Travis has internal ticket to track this feature request.