User function frame_order.decompose: Implementation of the 'total', 'reverse' and 'mirror' params.
This allows a fixed number of structures to be generated over the distribution, for the model order
to be reversed, and for the models to step from the negative angle to positive angle and then return
to the negative angle.

The original code has been simplified by switching from numpy.arange() to numpy.linspace() for
generating the range of angles.  This function is far more reliable than arange() which has end
point instability issues!