RelaxErrors: Expansion of the functionality of the BaseArgError base class.
The docstring now documents the arguments.  The 'dim' and 'rank' arguments have been added to allow
for more control over the reported message for array-type objects.  And the 'can_be_none' argument
has been added to append ', or None' to the message, negating the need for the RelaxNone*Error
objects.

For formatting the lists used in the BaseArgError class, the new function human_readable_list() has
been added to the lib.text.string module.