Python tempfile.mktemp(): Converted all usage of the function to tempfile.mkstemp().
The tempfile.mktemp() function was depreciated in Python 2.3.  According to the Python
documentation:

"""
A historical way to create temporary files was to first generate a file name with the mktemp()
function and then create a file using this name. Unfortunately this is not secure, because a
different process may create a file with this name in the time between the call to mktemp() and the
subsequent attempt to create the file by the first process. The solution is to combine the two steps
and create the file immediately. This approach is used by mkstemp() and the other functions
described above.
"""

The Travis CI testing system was sometimes failing on files created with mktemp(), so hopefully
mkstemp() will alleviate the issue.