Unit tests: Addition of two tests for specific_analyses.relax_disp.parameters.loop_parameters().
The two unit tests are Test_parameters.test_loop_parameters_clustered_spins and
Test_parameters.test_loop_parameters_single_spin in the unit test module
_specific_analyses._relax_disp.test_parameters.

These were added to try to catch the typo error at the end of the function, where the 'dwH_AB'
parameter appears twice (the second should be 'dwH_AC').  However the typo was not caught in the
tests as no currently implemented dispersion model contains the 'dwH_AC' parameter.  Hence it is a
latent bug.

The tests do catch a minor error with the 'R2eff' model in which the 'i0' parameter is always
returned.  'i0' should only be returned when exponential curve data is present.  This bug has no
apparent affect on the current operation of relax, so the parameter is probably handled correctly
downstream.