Dispersion: Fix for bug #8, the accidental modification of the hardcoded variables.
See https://sourceforge.net/p/nmr-relax/tickets/8/.  The MODEL_PARAMS lists in
specific_analysis.relax_disp.variables were accidentally being modified by the Model_class class in
the specific_analyses.relax_disp.model module.  The list for a given model was being set as the
self.params list.  This list would then have the 'r1' parameter pre-pended to it if that parameter
is optimised for a model, and hence the specific_analysis.relax_disp.variables list would be
permanently modified.  Now copy.deepcopy() is being used for all variables to avoid this issue.

This bug was uncovered in the unit tests as the _specific_analyses._relax_disp.test_model tests were
causing 'r1' to be added, and then the later _specific_analyses._relax_disp.test_parameters tests
would fail as 'r1' should not be in those lists.

This bug is highly unlikely to be encountered by users of relax.  You would need to run two
analyses, one after the other without closing relax, and the first analysis would need to have 'r1'
optimised and the second not.