Unit tests: Creation of the Test_parameters.test_param_conversion_clustered_spins_sim test.
This is to check the specific_analyses.relax_disp.parameters.param_conversion() function for a
cluster of 2 spins for Monte Carlo simulations.  It was a failed attempt to catch bug #10
(https://sourceforge.net/p/nmr-relax/tickets/10/).  The problem probably lies in the Monte Carlo
simulation setup functions in the specific analysis API rather than in the module
specific_analyses.relax_disp.parameters.