relax GUI: wxPython-Phoenix 4.x fix to allow relax to start again.
In the later wxPython versions, relax would not be able to start either the GUI or any of the test
suite due to a new error "wx._core.PyNoAppError: The wx.App object must be created first!".  This
was not present in wxPython-Phoenix 3.  The Relax_icons class (a wx.IconBundle derived class) is no
longer instantiated on import.