Structure loading: Fix for bug #12, the acceptance of the invalid '#' character in molecule names.
See https://sourceforge.net/p/nmr-relax/tickets/12/.  A simple check has not been added to the
load_pdb() and load_xyz() functions of the internal structural object in
lib.structure.internal.object.  This ensures that the '#' character can never be set as the molecule
name, independently if it was taken from a file name or set via the set_mol_name arguments of the
structure.read_pdb or structure.read_xyz user functions.