Fix for bug #14, the freezing of user functions in the GUI.
See https://sourceforge.net/p/nmr-relax/tickets/14/.  The user functions freeze if an error occurs
that is not a RelaxError, with the mouse pointer stuck on the busy cursor.  These non-RelaxErrors are
now caught and manually dealt with by the GUI interpreter.

Like all GUI freezing bugs, this was introduced with the huge GUI speed ups of
1321b95694f6d30174ae09406f167d1cf82b1961.  These also only to appear to be a freeze, but it is
actually the failure to update and show the relax controller combined with not turning off the busy
mouse cursor.