Relaxation dispersion: Protection of all of the MODEL_PARAMS_* variables from modification.
These are now only used with copy.deepcopy().  This removes a number of bugs in which the lists,
which should be read-only, are permanently modified by the addition of 'r1'.  The system tests add
'r1' and then the unit tests subsequently fail.  This would also be an issue if an experiment
without the 'r1' parameter is analysed after one with that parameter, without restarting relax.