FSF Copyright Validation script: Support for saving and reading the committer information.
This allows the committer information (file name, committer name, and copyright years) from older
repositories to be saved and later read into the script.  In this case, the old Subversion history
has been read and the committer information placed into the fsfcv.svn_committer_info.bz2 file (in
the devel_scripts/ directory).  This compressed file is now specified in the fsfcv.conf.py
configuration file.  The result is that the fsfcv script can be run on the relax git repository
without requiring a checkout of the old SVN repository.