Tavis CI config: Fix for the Mac OS X build.
This job passes, but the test suite fails with the following traceback message when trying to import
matplotlib:

"""
ImportError: Python is not installed as a framework. The Mac OS X backend will not be able to
function correctly if Python is not installed as a framework. See the Python documentation for more
information on installing Python as a framework on Mac OS X. Please either reinstall Python as a
framework, or try one of the other backends. If you are using (Ana)Conda please install python.app
and replace the use of 'python' with 'pythonw'. See 'Working with Matplotlib on OSX' in the
Matplotlib FAQ for more information.
"""

The fix is simply to create $HOME/.matplotlib/matplotlibrc with the contents 'backend: TkAgg'.