Travis CI config: Creation of a job for testing relax on an arm64 CPU.
The system Python and its packages are used to avoid timeouts on arm64.  Installing the Python
packages via pip prior to running causes a Travis CI time out, as most of the 50 minutes allowed are
used up by the compilation of SciPy.

Despite the successful installation of the wxPython site-package on the system Python3, the GUI
tests are not activated as there is a problem with xvfb on the arm64 Travis CI jobs.