Test suite: Implementation of a command line option for disabling IO capture.
This was previously handled by using the debug command line option which simply prevented IO
capture.  This type of output is very hard to parse by eye, as the tests are not well separated and
the debugging output is very verbose.

Now the --no-capt or --no-capture option has been implement to disable the IO capture.  The debug
command line option no longer disables IO capture, rather it allows for finer control of the test
suite in that verbose debugging output is now only shown for tests that do not pass.

When IO capture is disabled, extra formatted output is used to provide clear separators, titles,
descriptions and endings for each test.